intermediate_level/fast_api/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "YaronEsquivel",email = "yaron.esquivel@axity.com"}
]
requires-python = ">=3.12"
dependencies = [
    "black (>=26.3.1,<27.0.0)",
    "ruff (>=0.15.10,<0.16.0)",
    "isort (>=8.0.1,<9.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "pydantic[email] (>=2.13.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
